refactor in code structure -> separation
