+++ README.md
@@ -4,4 +4,5 @@
 
 #### Thank you for your cooperation in advance!
 
-*20 March 2017*
+######dsztanko
+*20-21 March 2017*
